+++ client/src/jvmTest/kotlin/eu/tiducto/spider/client/architecture/OpenApiPinTest.kt
@@ -8,6 +8,9 @@ import eu.tiducto.spider.contract.realtime.ActivePeriodDto
 import eu.tiducto.spider.contract.realtime.AlertDto
 import eu.tiducto.spider.contract.realtime.AlertsResponseDto
 import eu.tiducto.spider.contract.realtime.DelayDto
+import eu.tiducto.spider.contract.realtime.DelayGroupResultDto
+import eu.tiducto.spider.contract.realtime.DelayQueryDto
+import eu.tiducto.spider.contract.realtime.DelaysRequestDto
 import eu.tiducto.spider.contract.realtime.DelaysResponseDto
 import eu.tiducto.spider.contract.realtime.InformedEntityDto
 import eu.tiducto.spider.contract.realtime.StopTimeUpdateDto
@@ -72,7 +75,10 @@ class OpenApiPinTest {
             "VehiclesResponse" to VehiclesResponseDto.serializer().descriptor,
             "VehicleByTripResponse" to VehicleByTripResponseDto.serializer().descriptor,
             "Vehicle" to VehicleDto.serializer().descriptor,
+            "DelaysRequest" to DelaysRequestDto.serializer().descriptor,
+            "DelayQuery" to DelayQueryDto.serializer().descriptor,
             "DelaysResponse" to DelaysResponseDto.serializer().descriptor,
+            "DelayGroupResult" to DelayGroupResultDto.serializer().descriptor,
             "Delay" to DelayDto.serializer().descriptor,
             "StopTimeUpdate" to StopTimeUpdateDto.serializer().descriptor,
             "AlertsResponse" to AlertsResponseDto.serializer().descriptor,